Replaced 'double' by 'double precision' for compatibility
git-svn-id: http://svn.transientskp.org/code/tkp/trunk@2279 2b73c8c1-3922-0410-90dd-bc0a5c6f2ac6

+++ tkp/database/utils.py
@@ -1574,8 +1574,8 @@ WHERE
       %(catalog_filter)s
   cs.cat_id = c.catid
   AND cs.x * %%s + cs.y * %%s + cs.z * %%s > COS(RADIANS(%%s))
-  AND cs.zone BETWEEN CAST(FLOOR((CAST(%%s AS DOUBLE) - %%s) / %%s) AS INTEGER)
-                  AND CAST(FLOOR((CAST(%%s AS DOUBLE) + %%s) / %%s) AS INTEGER)
+  AND cs.zone BETWEEN CAST(FLOOR((CAST(%%s AS DOUBLE precision) - %%s) / %%s) AS INTEGER)
+                  AND CAST(FLOOR((CAST(%%s AS DOUBLE precision) + %%s) / %%s) AS INTEGER)
   AND cs.ra BETWEEN %%s - alpha(%%s, %%s)
                 AND %%s + alpha(%%s, %%s)
   AND cs.decl BETWEEN %%s - %%s
